+++ dbt/include/spark/macros/materializations/incremental/strategies.sql
@@ -20,7 +20,8 @@
 
 
 {% macro spark__get_merge_sql(target, source, unique_key, dest_columns, predicates=none) %}
-  {# ignore dest_columns - we will just use `*` #}
+  {# skip dest_columns, use merge_update_columns config if provided, otherwise use "*" #}
+  {%- set update_columns = config.get("merge_update_columns") -%}
   
   {% set merge_condition %}
     {% if unique_key %}
@@ -32,8 +33,16 @@
   
     merge into {{ target }} as DBT_INTERNAL_DEST
       using {{ source.include(schema=false) }} as DBT_INTERNAL_SOURCE
+      
       {{ merge_condition }}
-      when matched then update set *
+      
+      when matched then update set
+        {% if update_columns -%}{%- for column_name in update_columns %}
+            {{ column_name }} = DBT_INTERNAL_SOURCE.{{ column_name }}
+            {%- if not loop.last %}, {%- endif %}
+        {%- endfor %}
+        {%- else %} * {% endif %}
+    
       when not matched then insert *
 {% endmacro %}